% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_tables.R
\name{azure_table}
\alias{azure_table}
\alias{azure_table.table_endpoint}
\alias{list_azure_tables}
\alias{list_azure_tables.table_endpoint}
\alias{create_azure_table}
\alias{create_azure_table.table_endpoint}
\alias{delete_azure_table}
\alias{delete_azure_table.table_endpoint}
\alias{delete_azure_table.azure_table}
\title{Operations with azure tables}
\usage{
azure_table(endpoint, ...)

\method{azure_table}{table_endpoint}(endpoint, name, ...)

list_azure_tables(endpoint, ...)

\method{list_azure_tables}{table_endpoint}(endpoint, ...)

create_azure_table(endpoint, ...)

\method{create_azure_table}{table_endpoint}(endpoint, name, ...)

delete_azure_table(endpoint, ...)

\method{delete_azure_table}{table_endpoint}(endpoint, name, confirm = TRUE, ...)

\method{delete_azure_table}{azure_table}(endpoint, ...)
}
\arguments{
\item{endpoint}{An object of class \code{table_endpoint}.}

\item{...}{Other arguments passed to lower-level functions.}

\item{name}{The name of a table in a storage account.}

\item{confirm}{For deleting a table, whether to ask for confirmation.}
}
\description{
Operations with azure tables
}
\details{
These methods are for accessing and managing tables within a storage account.
}
\seealso{
\link{table_endpoint}, \link{table_entity}
}
