% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attachment.R
\name{list_attachments}
\alias{list_attachments}
\alias{cosmos_attachment}
\alias{list_attachments.cosmos_document}
\alias{create_attachment}
\alias{create_attachment.cosmos_document}
\alias{download_attachment}
\alias{download_attachment.cosmos_attachment}
\alias{delete_attachment}
\alias{delete_attachment.cosmos_document}
\alias{delete_attachment.cosmos_attachment}
\title{Methods for working with Azure Cosmos DB attachments}
\usage{
list_attachments(document, ...)

\method{list_attachments}{cosmos_document}(document, ...)

create_attachment(document, ...)

\method{create_attachment}{cosmos_document}(
  document,
  file,
  content_type = "application/octet-stream",
  id = NULL,
  headers = list(),
  ...
)

download_attachment(attachment, ...)

\method{download_attachment}{cosmos_attachment}(
  attachment,
  destfile,
  options = list(),
  headers = list(),
  overwrite = FALSE,
  ...
)

delete_attachment(document, ...)

\method{delete_attachment}{cosmos_document}(document, id, confirm = TRUE, ...)

\method{delete_attachment}{cosmos_attachment}(document, ...)
}
\arguments{
\item{document}{A Cosmos DB document object, as obtained by \code{get_document}, \code{create_document}, \code{list_documents} or \code{query_documents}.}

\item{file}{For \code{create_attachment}, the file to turn into an attachment. This can be a filename, a raw connection, or a URL connection pointing to a location on the Internet.}

\item{content_type}{For \code{create_attachment}, the content type of the attachment. Defaults to "application/octet-stream".}

\item{id}{For \code{create_attachment} and \code{delete_attachment}, the ID for the attachment. This is optional for \code{create_attachment}.}

\item{attachment}{For \code{download_attachment}, the attachment object, as obtained by \code{list_attachments} or \code{create_attachment}.}

\item{destfile}{For \code{download_attachment}, the destination filename.}

\item{options, headers, ...}{Optional arguments passed to lower-level functions.}

\item{overwrite}{For \code{download_attachment}, whether to overwrite an existing destination file.}

\item{confirm}{For \code{delete_attachment}, whether to ask for confirmation before deleting.}
}
\description{
Methods for working with Azure Cosmos DB attachments
}
\details{
These are methods for working with document attachments in Cosmos DB using the core (SQL) API. Note that the attachments API is deprecated; going forward, you should explore other options for hosting external files, such as Azure blob storage.
}
