% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{get_document}
\alias{get_document}
\alias{cosmos_document}
\alias{create_document}
\alias{create_document.cosmos_container}
\alias{list_documents}
\alias{list_documents.cosmos_container}
\alias{delete_document}
\alias{delete_document.cosmos_container}
\alias{delete_document.cosmos_document}
\title{Methods for working with Azure Cosmos DB documents}
\usage{
get_document(container, ...)

create_document(container, ...)

\method{create_document}{cosmos_container}(container, data, headers = list(), ...)

list_documents(container, ...)

\method{list_documents}{cosmos_container}(
  container,
  partition_key = NULL,
  as_data_frame = FALSE,
  metadata = TRUE,
  headers = list(),
  ...
)

delete_document(container, ...)

\method{delete_document}{cosmos_container}(
  container,
  id,
  partition_key,
  headers = list(),
  confirm = TRUE,
  ...
)

\method{delete_document}{cosmos_document}(container, ...)
}
\arguments{
\item{container}{A Cosmos DB container object, as obtained by \code{get_cosmos_container} or \code{create_cosmos_container}.}

\item{data}{For \code{create_document}, the document data. This can be either a string containing JSON text, or a (possibly nested) list containing the parsed JSON.}

\item{headers, ...}{Optional arguments passed to lower-level functions.}

\item{partition_key}{For \code{get_document} and \code{delete_document}, the value of the partition key for the desired document. For \code{list_documents}, restrict the returned list only to documents with this key value.}

\item{as_data_frame}{For \code{list_documents}, whether to return a data frame or a list of Cosmos DB document objects.}

\item{metadata}{For \code{get_document} and \code{list_documents}, whether to include Cosmos DB document metadata in the result.}

\item{id}{The document ID.}

\item{confirm}{For \code{delete_cosmos_container}, whether to ask for confirmation before deleting.}
}
\description{
Methods for working with Azure Cosmos DB documents
}
\details{
These are low-level functions for working with individual documents in a Cosmos DB container. In most cases you will want to use \link{query_documents} to issue queries against the container, or \link{bulk_import} and \link{bulk_delete} to create and delete documents.
}
\seealso{
\link{query_documents}, \link{bulk_import}, \link{bulk_delete}, \link{cosmos_container}
}
