% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongo_endpoint.R
\name{cosmos_mongo_endpoint}
\alias{cosmos_mongo_endpoint}
\alias{cosmos_mongo_connection}
\alias{cosmos_mongo_connection.cosmos_mongo_endpoint}
\title{MongoDB endpoint for Azure Cosmos DB}
\usage{
cosmos_mongo_endpoint(
  host,
  key,
  mongo_options = list(),
  connection_string = NULL
)

cosmos_mongo_connection(endpoint, ...)

\method{cosmos_mongo_connection}{cosmos_mongo_endpoint}(endpoint, collection, database, ...)
}
\arguments{
\item{host}{for \code{cosmos_mongo_endpoint}, the URL of the Cosmos DB MongoDB endpoint. Usually of the form "https://{account-name}.mongo.cosmos.azure.com:443/".}

\item{key}{A string containing the access key (password) for the endpoint. Can be either a read-write or read-only key.}

\item{mongo_options}{A named list containing any additional parameters for the Mongo DB connection string.}

\item{connection_string}{Optionally, the full connection string for the MongoDB endpoint. If this is supplied, all other arguments are ignored.}

\item{endpoint}{For \code{cosmos_mongo_connection}, a MongoDB endpoint object as obtained from \code{cosmos_mongo_endpoint}.}

\item{...}{Optional arguments passed to lower-level functions.}

\item{collection, database}{For \code{cosmos_mongo_connection}, the collection and database to connect to.}
}
\value{
For \code{cosmos_mongo_endpoint}, an object of S3 class \code{cosmos_mongo_endpoint}.

For \code{cosmos_mongo_connection}, an object of class \code{mongolite::mongo} which can then be used to interact with the given collection.
}
\description{
MongoDB endpoint for Azure Cosmos DB
}
\details{
These functions are the basis of the AzureCosmosR MongoDB API client. They serve as a bridge between the Azure endpoint and the MongoDB connection functionality provided by the mongolite package.
}
\seealso{
\link{az_cosmosdb}, \link[mongolite:mongo]{mongolite::mongo}

For the SQL API client framework: \link{cosmos_endpoint}, \link{cosmos_database}, \link{cosmos_container}, \link{query_documents}
}
