% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_methods.R
\name{delete_cosmosdb_account}
\alias{delete_cosmosdb_account}
\title{Delete Azure Cosmos DB account}
\description{
Method for the \link[AzureRMR:az_resource_group]{AzureRMR::az_resource_group} class.
}
\section{Usage}{
\preformatted{delete_cosmosdb_account(name, confirm = TRUE, wait = FALSE)
}
}

\section{Arguments}{

\itemize{
\item \code{name}: The name of the Cosmos DB account.
\item \code{confirm}: Whether to ask for confirmation before deleting.
\item \code{wait}: Whether to wait until the deletion has completed before returning.
}
}

\section{Details}{

This method deletes an existing Azure Cosmos DB account.
}

\seealso{
\link{create_cosmosdb_account}, \link{get_cosmosdb_account}

For the SQL API client framework: \link{cosmos_endpoint}, \link{cosmos_database}, \link{cosmos_container}, \link{query_documents}

For the table storage API: \link[AzureTableStor:table_endpoint]{AzureTableStor::table_endpoint}

For the MongoDB API: \link{cosmos_mongo_endpoint}, \link[mongolite:mongo]{mongolite::mongo}
}
