% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_key.R
\name{get_partition_key}
\alias{get_partition_key}
\alias{list_partition_key_values}
\alias{list_partition_key_ranges}
\title{Container partition key information}
\usage{
get_partition_key(container)

list_partition_key_values(container)

list_partition_key_ranges(container)
}
\arguments{
\item{container}{An object of class \code{cosmos_container}.}
}
\value{
For \code{get_partition_key}, the name of the partition key column as a string.

For \code{list_partition_key_values}, a character vector of all the values of the partition key.

For \code{list_partition_key_ranges}, a character vector of the IDs of the partition key ranges.
}
\description{
Container partition key information
}
\details{
These are functions to facilitate working with a Cosmos DB container, which often requires knowledge of its partition key.
}
