% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_endpoint.R
\name{table_endpoint}
\alias{table_endpoint}
\alias{call_table_endpoint}
\title{Table storage endpoint}
\usage{
table_endpoint(
  endpoint,
  key = NULL,
  token = NULL,
  sas = NULL,
  api_version = getOption("azure_storage_api_version")
)

call_table_endpoint(
  endpoint,
  path,
  options = list(),
  headers = list(),
  body = NULL,
  ...,
  metadata = c("none", "minimal", "full")
)
}
\arguments{
\item{endpoint}{For \code{table_endpoint}, the URL (hostname) of the table storage endpoint: for the Azure public cloud, this will be of the form \verb{https://\{account-name\}.table.core.windows.net}. For \code{call_table_endpoint}, an object of class \code{table_endpoint}.}

\item{key}{The access key for the storage account.}

\item{token}{An Azure Active Directory (AAD) authentication token. Not used for table storage.}

\item{sas}{A shared access signature (SAS) for the account.}

\item{api_version}{The storage API version to use when interacting with the host. Defaults to "2019-07-07".}

\item{path}{For \code{call_table_endpoint}, the path component of the endpoint call.}

\item{options}{For \code{call_table_endpoint}, a named list giving the query parameters for the operation.}

\item{headers}{For \code{call_table_endpoint}, a named list giving any additional HTTP headers to send to the host. AzureCosmosR will handle authentication details, so you don't have to specify these here.}

\item{body}{For \code{call_table_endpoint}, the request body for a PUT/POST/PATCH call.}

\item{...}{For \code{call_table_endpoint}, further arguments passed to \code{AzureStor::call_storage_endpoint} and \code{httr::VERB}.}

\item{metadata}{For \code{call_table_endpoint}, the level of ODATA metadata to include in the response.}
}
\description{
Table storage endpoint object, and method to call it.
}
