% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\name{get_cosmos_container}
\alias{get_cosmos_container}
\alias{cosmos_container}
\alias{get_cosmos_container.cosmos_database}
\alias{create_cosmos_container}
\alias{create_cosmos_container.cosmos_database}
\alias{delete_cosmos_container}
\alias{delete_cosmos_container.cosmos_database}
\alias{delete_cosmos_container.cosmos_container}
\alias{list_cosmos_containers}
\alias{list_cosmos_containers.cosmos_database}
\alias{get_partition_key}
\alias{get_partition_key.cosmos_container}
\alias{list_key_values}
\alias{list_key_values.cosmos_container}
\title{Methods for working with Cosmos DB containers (tables)}
\usage{
get_cosmos_container(database, ...)

\method{get_cosmos_container}{cosmos_database}(database, name, ...)

create_cosmos_container(database, ...)

\method{create_cosmos_container}{cosmos_database}(
  database,
  name,
  partition_key,
  partition_version = 2,
  autoscale_maxRUs = NULL,
  manual_RUs = NULL,
  headers = list(),
  ...
)

delete_cosmos_container(database, ...)

\method{delete_cosmos_container}{cosmos_database}(database, name, confirm = TRUE, ...)

\method{delete_cosmos_container}{cosmos_container}(database, ...)

list_cosmos_containers(database, ...)

\method{list_cosmos_containers}{cosmos_database}(database, ...)

get_partition_key(container)

\method{get_partition_key}{cosmos_container}(container)

list_key_values(container)

\method{list_key_values}{cosmos_container}(container)
}
\arguments{
\item{database}{A Cosmos DB database object, as obtained from \code{get_cosmos_database} or \code{create_cosmos_database}, or for \code{delete_cosmos_container.cosmos_container}, the container object.}

\item{...}{Arguments passed to lower-level functions.}

\item{name}{The name of the container.}

\item{partition_key}{For \code{create_cosmos_container}, the name of the partition key.}

\item{partition_version}{For \code{create_cosmos_container}, the partition version. Can be either 1 or 2. Version 2 supports large partition key values (longer than 100 bytes) but requires API version \code{2018-12-31} or later. Use version 1 if the container needs to be accessible to older Cosmos DB SDKs.}

\item{autoscale_maxRUs, manual_RUs}{For \code{create_cosmos_container}, optional parameters for the maximum request units (RUs) allowed. See the Cosmos DB documentation for more details.}

\item{headers}{For \code{create_cosmos_container}, optional HTTP headers to include in the request.}

\item{confirm}{For \code{delete_cosmos_container}, whether to ask for confirmation before deleting.}
}
\description{
Methods for working with Cosmos DB containers (tables)
}
\details{
These are methods for working with Cosmos DB containers (tables) using the core (SQL) API.

\code{get_cosmos_container}, \code{create_cosmos_container}, \code{delete_cosmos_container} and \code{list_cosmos_containers} provide basic container management functionality.

\code{get_partition_key} returns the name of the partition key column in the container, and \code{list_key_values} returns a vector of the distinct values for this column. These are useful when working with queries that have to be mapped across partitions.
}
\seealso{
\link{cosmos_container}, \link{query_documents}
}
