% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sproc.R
\name{list_stored_procedures}
\alias{list_stored_procedures}
\alias{cosmos_stored_procedure}
\alias{create_stored_procedure}
\alias{create_stored_procedure.cosmos_container}
\alias{exec_stored_procedure}
\alias{exec_stored_procedure.cosmos_container}
\alias{exec_stored_procedure.cosmos_stored_procedure}
\alias{replace_stored_procedure}
\alias{replace_stored_procedure.cosmos_container}
\alias{replace_stored_procedure.cosmos_stored_procedure}
\alias{delete_stored_procedure}
\alias{delete_stored_procedure.cosmos_container}
\alias{delete_stored_procedure.cosmos_stored_procedure}
\title{Methods for working with Azure Cosmos DB stored procedures}
\usage{
list_stored_procedures(object, ...)

create_stored_procedure(object, ...)

\method{create_stored_procedure}{cosmos_container}(object, procname, body, ...)

exec_stored_procedure(object, ...)

\method{exec_stored_procedure}{cosmos_container}(object, procname, parameters = list(), ...)

\method{exec_stored_procedure}{cosmos_stored_procedure}(object, ...)

replace_stored_procedure(object, ...)

\method{replace_stored_procedure}{cosmos_container}(object, procname, body, ...)

\method{replace_stored_procedure}{cosmos_stored_procedure}(object, body, ...)

delete_stored_procedure(object, ...)

\method{delete_stored_procedure}{cosmos_container}(object, procname, confirm = TRUE, ...)

\method{delete_stored_procedure}{cosmos_stored_procedure}(object, ...)
}
\arguments{
\item{object}{A Cosmos DB container object, as obtained by \code{get_cosmos_container} or \code{create_cosmos_container}, or for `delete_stored_procedure.cosmos_stored_procedure, the stored procedure object.}

\item{...}{Optional arguments passed to lower-level functions.}

\item{procname}{The name of the stored procedure.}

\item{body}{For \code{create_stored_procedure} and \code{replace_stored_procedure}, the body of the stored procedure as text.}

\item{parameters}{For \code{exec_stored_procedure}, a list of parameters to pass to the procedure.}

\item{confirm}{For \code{delete_cosmos_container}, whether to ask for confirmation before deleting.}
}
\description{
Methods for working with Azure Cosmos DB stored procedures
}
\details{
These are methods for working with stored procedures in Azure Cosmos DB using the core (SQL) API. In the Cosmos DB model, stored procedures are written in JavaScript and associated with a container.
}
\examples{
\dontrun{

# example text of a stored procedure: uploading multiple rows of data
readLines(system.file("srcjs/bulkUpload.js", package="AzureCosmosR"))

endp <- cosmos_endpoint("https://myaccount.documents.azure.com:443/", key="mykey")
db <- get_cosmos_database(endp, "mydatabase")
cont <- create_cosmos_container(db, "mycontainer", partition_key="sex")

proc <- create_stored_procedure(cont, "myBulkUpload",
    body=readLines(system.file("srcjs/bulkUpload.js", package="AzureCosmosR")))

list_stored_procedures(cont)

sw_male <- dplyr::filter(dplyr::starwars, sex == "male")
exec_stored_procedure(proc, parameters=list(sw_male))

delete_stored_procedure(proc)

}
}
